+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - HTML parser error with single closing square bracket in table row #1893
 - reference counted strings put the HTML bit in the middle of the reference
   count #1984
+- &amp;amp; escape disappearing #797
 
 ## [2.47.0] - 2021-03-15